refactor(pdf): isolate equation reconstruction
